pyproject.toml
[tool.poetry]
name = "Weather"
version = "0.1.0"
description = ""
authors = ["sorelsh <sorelsh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
seaborn = "^0.13.2"
pandas = "^2.2.3"
streamlit = "^1.41.1"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
